packages/ai-services/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="JACAMENO AI Services", version="1.0.0")

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Models
class MixMasterRequest(BaseModel):
    tracks: List[Dict[str, Any]]
    preferences: Optional[Dict[str, Any]] = None

class SongwritingRequest(BaseModel):
    prompt: str
    genre: Optional[str] = None
    mood: Optional[str] = None
    structure: Optional[List[str]] = None

class VocalCoachRequest(BaseModel):
    audioUrl: str
    targetStyle: Optional[str] = None

# Health check
@app.get("/health")
async def health_check():
    return {"status": "ok", "service": "JACAMENO AI Services"}

# AI Mixing & Mastering
@app.post("/api/mix-master")
async def mix_master(request: MixMasterRequest):
    """
    AI-powered mixing and mastering service
    Analyzes tracks and applies professional mixing/mastering
    """
    try:
        # Initialize default preferences
        prefs = request.preferences or {}
        genre = prefs.get('genre', 'auto')
        intensity = prefs.get('intensity', 'medium')
        target_loudness = prefs.get('targetLoudness', -14)
        
        # AI processing would happen here
        # - Analyze frequency spectrum
        # - Apply EQ adjustments
        # - Compression and limiting
        # - Spatial enhancement
        # - Loudness normalization
        
        result = {
            "status": "success",
            "processedTracks": len(request.tracks),
            "settings": {
                "genre": genre,
                "intensity": intensity,
                "targetLoudness": target_loudness,
                "eqApplied": True,
                "compressionApplied": True,
                "limitingApplied": True,
                "spatialEnhancement": True
            },
            "analysis": {
                "detectedGenre": "auto-detected",
                "averageLoudness": -12.5,
                "dynamicRange": 8.2,
                "frequencyBalance": "optimal"
            }
        }
        
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# AI Songwriting Assistant
@app.post("/api/songwriting")
async def songwriting_assistant(request: SongwritingRequest):
    """
    AI songwriting assistant using OpenAI
    Generates lyrics, chord progressions, and melody suggestions
    """
    try:
        # Would use OpenAI API here
        suggestions = {
            "lyrics": {
                "verse1": "AI-generated verse 1...",
                "chorus": "AI-generated chorus...",
                "verse2": "AI-generated verse 2...",
                "bridge": "AI-generated bridge..."
            },
            "chordProgressions": [
                {
                    "section": "verse",
                    "chords": ["C", "Am", "F", "G"],
                    "pattern": "4/4"
                },
                {
                    "section": "chorus",
                    "chords": ["F", "C", "G", "Am"],
                    "pattern": "4/4"
                }
            ],
            "melodyNotes": {
                "verse": ["C4", "D4", "E4", "G4"],
                "chorus": ["E4", "G4", "A4", "C5"]
            },
            "structure": request.structure or ["verse", "chorus", "verse", "chorus", "bridge", "chorus"],
            "genre": request.genre or "pop",
            "mood": request.mood or "uplifting"
        }
        
        return suggestions
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# AI Vocal Coaching
@app.post("/api/vocal-coach")
async def vocal_coach(request: VocalCoachRequest):
    """
    AI vocal coaching service
    Analyzes vocal performance and provides feedback
    """
    try:
        # Would perform audio analysis here
        # - Pitch detection and accuracy
        # - Timing and rhythm analysis
        # - Tone quality assessment
        # - Breath control evaluation
        
        feedback = {
            "overallScore": 7.5,
            "pitchAccuracy": {
                "score": 8.0,
                "feedback": "Generally good pitch control. Watch for slight sharp tendency on high notes."
            },
            "timing": {
                "score": 7.5,
                "feedback": "Rhythmic accuracy is good. Consider tightening timing on fast passages."
            },
            "tone": {
                "score": 7.0,
                "feedback": "Nice tone quality. Focus on consistent breath support for sustained notes."
            },
            "improvements": [
                {
                    "area": "High Notes",
                    "suggestion": "Relax jaw and open throat for clearer high notes",
                    "exercises": ["Lip trills", "Sirens"]
                },
                {
                    "area": "Breath Support",
                    "suggestion": "Practice diaphragmatic breathing",
                    "exercises": ["Hissing exercise", "Sustained 'ah' sounds"]
                }
            ],
            "targetStyle": request.targetStyle,
            "comparisonMetrics": {
                "pitchRange": "2.5 octaves",
                "dynamicRange": "Good",
                "vibrato": "Natural"
            }
        }
        
        return feedback
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Audio Analysis
@app.post("/api/analyze-audio")
async def analyze_audio(audioUrl: str):
    """
    General audio analysis service
    """
    try:
        analysis = {
            "duration": 180.5,
            "bpm": 120,
            "key": "C major",
            "timeSignature": "4/4",
            "averageLoudness": -14.2,
            "peakLoudness": -3.5,
            "dynamicRange": 9.0,
            "frequencySpectrum": {
                "bass": "balanced",
                "mids": "present",
                "highs": "bright"
            }
        }
        
        return analysis
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
